fix: keep reminder delivery compatible with bare Node
