+++ scripts/lib/security-gate.test.mjs
@@ -1,8 +1,7 @@
-import "./ts-esm-hooks.mjs";
 import assert from "node:assert/strict";
 import test from "node:test";
 
-const { sanitizeInbound } = await import("../../agent/lib/security-gate.ts");
+const { sanitizeInbound } = await import("./security-gate.mjs");
 
 test("sanitizeInbound reports exact Unicode code points removed at N-1, N and N+1", () => {
   const nMinusOne = sanitizeInbound("🙂".repeat(2), 3);
