+++ scripts/lib/telegram-send.mjs
@@ -12,7 +12,7 @@
 // htmlToPlain (HTML→plain с декодом сущностей) живёт в общем модуле — тот же
 // фолбэк-декодер использует и Telegram-канал (agent/channels/telegram.ts).
 import { toTelegramHtmlChunks, htmlToPlain } from "./telegram-format.mjs";
-import { scanOutbound } from "../../agent/lib/security-gate.ts";
+import { scanOutbound } from "./security-gate.mjs";
 
 async function post(bot, body) {
   const res = await fetch(`https://api.telegram.org/bot${bot}/sendMessage`, {
